Use 128 bit hash for graph construction.
Large potential for collision with small k and l parameters.
To fix this we can spread out the hash for small k's but this
will result in a slower hash implementation.

Also removed identifying k-min-mers with their reverse.
If this is something we really want we can implement it by
using two 128 bit hashes, one for regular and one for reverse.

Added -c/--check-collisions command line parameter to calculate
and display number of nodes that have collisions.